requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.6
Flask-WTF==1.1.2
